requirements.txt
flet
torch
accelerate
openai-whisper
pyaudio
wave
ffmpeg-python
XFormers
llama-cpp-python 
requests
desktop-notifier